feat : Application에서 객체 변환하던 역할 담당 Convertor 클래스 생성
